refactor(workflow): split runtime store facade
